works for me now after a lot of fixes, also working on removing old way of subtitles
