+++ utils/.config.template.toml
@@ -203,7 +203,6 @@ resolution_w = { optional = false, default = 1080, example = 1440, explantation 
 resolution_h = { optional = false, default = 1920, example = 2560, explantation = "Sets the height in pixels of the final video" }
 zoom = { optional = true, default = 1, example = 1.1, explanation = "Sets the browser zoom level. Useful if you want the text larger.", type = "float", nmin = 0.1, nmax = 2, oob_error = "The text is really difficult to read at a zoom level higher than 2" }
 channel_name = { optional = true, default = "Reddit Tales", example = "Reddit Stories", explanation = "Sets the username that appaers in the fancy title screen" }
-words_on_screen = { optional = true, default = 10, example = 1, type = "int", explanation = "Sets how many words should be shown on screen" }
 
 [settings.background]
 background_video = { optional = true, default = "minecraft", example = "rocket-league", options = [
